use of setter/getter when appropriates (closes GH-7)
This patch transform a few methods to properties, with getter and setter when
required.

A few tests where added and fixed.

Some docstring where also updated to reflect to specify if the method is in
fact a property.

Some refactoring of the Header class was also required.

Previously some methods where imported directly from the Header object to
Response and Request using `getattr'. This code is now removed and the
properties where added directly to the code.